backend/requirements.txt
# Web框架
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# 数据模型与配置
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 抖音API（内置X-Bogus/ABogus处理）
f2>=0.0.1.7

# 异步HTTP
httpx>=0.27.0

# 图片处理
pillow>=10.0.0
pillow-heif>=1.0.0

# 模板
jinja2>=3.1.0
aiofiles>=24.0.0
python-multipart>=0.0.9

# 配置解析
pyyaml>=6.0

# Windows 桌面客户端
pywebview>=6.0

# 打包（可选）
# pyinstaller>=6.0

# 系统通知（可选，桌面客户端用）
# plyer>=2.0
